requirements.txt
anki-vector
Pillow
flask
flask-sqlalchemy
flask-bcrypt
flask-login
flask-wtf
flask-migrate
flask-script
flask-restful
Flask-APScheduler
Flask-Avatars
htmlmin
jsmin
python-dateutil
beautifulsoup4
flask_caching
flask-socketio
gitpython